Buy and Sell Orders completed Transactions not working yet

+++ SuperTraders.Core/Entities/UserShare.cs
@@ -1,16 +1,12 @@
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
-
 namespace SuperTraders.Core.Entities
 {
-    public class UserShare: BaseEntity
+    public class UserShare : BaseEntity
     {
-        public float Amount { get; set; }
+        public int Amount { get; set; }
+        
         public string UserId { get; set; }
         public User User { get; set; }
-        
+
         public string ShareId { get; set; }
         public Share Share { get; set; }
     }
